only remove /live if it still exists

+++ initrd/live/custom/MX/8.sh
@@ -14,7 +14,7 @@ antix_specific_code() {
     rm -f $dir/etc/fstab.hotplug
     
     #remove leftover /live directory (after switch_root)
-    rm -R $dir/live
+    test -d "$dir/live" && rm -R $dir/live
     #symlink /live to FINAL_DIR location after switch_root
     ln -s /run/initramfs $dir/live
